Add project files
Create a setup.py file for the package that will use setuptools if
available, but will fall back to the lecagy standard library distutils
otherwise.

Create a make file to build various package format, install and remove
this package, and some other housekeeping.  The makefile recipes try to
use pip for installation/uninstallation if available, otherwise fall
back to the old style installation using python "egg" format.

GNU make extensions are avoided in the makefile to make this easier to
run on non-linux systems, which may not have GNU make installed or in
the path.

The makefile creates a version.txt file which contains the version
string for the package metadata (set in setup.py) and the package
__version__ data string (set in the top level __init__.py).

The Makefile.config file is created by the makefile which detects the
paths to pip, python, and pyflakes. This file can be changed to use
different values if desired, for example, to switch to python2 or
python3.