Add git-gerrit-unpicked command
Add a command to find commits and gerrit numbers of the commits on the
master branch, which have not been cherry picked on the stable branch.
Also can be used to show which commits are on the new stable branch but
not on the old stable branch, e.g, on 1.8.x, but not 1.6.x.

This is a rather heavy-weight command, but seems to do ok on a modern
system. (It requires on the order of 15MB to run, which is still about
20 times less than firefox in my system.)