Configurable --branch default for gerrit-fetch and gerrit-checkout
Use python string formatting to specify the gerrit-fetch and
gerrit-checkout branch names (much like the --format template for
output). This gives us a lot of flexibility on the branch names.

Support overriding the default 'gerrit/{number}/{patchset}' by setting
gerrit.fetchbranch in the git config. This can be handy if you don't
like to include the patchset in the branch or you just want a different
default prefix.