Rename gerrit-review to gerrit-update
Rename the git-gerrit-review command to git-gerrit-update, and rename
the git_gerrit.review() function to git_gerrit.update(), since this
command is used to update a gerrit change.

As it turns out, this command is commonly used to update a batch of
gerrits, so a common use to do git-gerrit-query and loop over the
changes to update each gerrit. So instead of taking only a single gerrit
number, have the new git-gerrit-update command take a set of search
terms and perform the query within the command, and update each gerrit
found. You can provide just the gerrit number as a search term, so this
is compatable with the old git-gerrit-review command.  Add a --dryrun
(-n) argument to just run the query.

Later, we can add a git-gerrit-review command to actually review code.

Unfortuately, this is a breaking change, so bump the version to
2.0.0rc1.