tests: Add a mocked query tests
Mock the git config and the pygerrit2 REST get to do a basic query test.
The test data is a sanitized version of data captured from a real query.

Add a smoke test for the query() function and the git_gerrit_query() cli
function.

Update the test for the git-gerrit-help cli subcommand. Capture the
output and check the first line is what we expect.

Add a test for the git-gerrit-version cli subcommand.