Add writeln() function
Currently, different funtions are used to write output to stdout and
stderr, making it harder to unit test.  Considiate all of these with a
new wrapper called writeln() to make it easier to mock in unit tests.

* Replace print() function calls with writeln()
* Replace the calls to print_info() and print_error() with the new
  writeln() helper function.
* Replace the calls to sys.stdout.write() and sys.stderr.write with
  the new writeln() helper function.
* Remove the from future statements for the print function, since we
  no longer use the print builtin function.

Mock the writeln() calls in the unit tests with the output fixture in
conftest.py. The fixture captures stdout and stderr during the tests and
so the tests can check the output.