Query more than 500 changes from Gerrit
The Gerrit REST API limits responses to 500 changes per query.  The
_more_changes is set to true on the last change if more changes are
available, and the caller can request the next batch of changes by
setting the start index with the 'S' option.

Implement this in the query generator by adding a do .. while loop to
retrieve the next batch of changes when the last change has the
_more_changes set to true.  Be sure to continue to support the limit
option correctly.