Add git-gerrit-sync
Add a local sqlite3 database in the project .git directory to store
gerrit numbers, patchset numbers, gerrit change-ids, and git
cherry-picked-from commit object ids.

Add the git-gerrit-sync command to fetch the entire set of changes
from the remote gerrit to the local repo and then scan the refs
and commit headers to populate the database.

The database is a local cache, so it should be safe to delete it and
recreate it at any time with git-gerrit-sync.

Query git-gerrit.db in gerrit-log

Query the database when running git-gerrit-log to quickly find the
gerrit numbers, gerrit change-ids, and cherry-picked-from gerrit
numbers.

Continue to scan the trailers to find the "Reviewed-by" trailer to
extract gerrit numbers, and use that in case the gerrit number is not in
the database, which can happen if the git-gerrit-sync was not run.

Remove the default "-" values from the changes returned by
git_gerrit.log(), which were a hack for formatting. Later we can add a
customized formatter to handle providing default values as part of the
format specifier.