Do not create branches by default
Do not create branches by default when running git-gerrit-fetch and
git-gerrit-checkout.  Originally, this seemed like a good idea, but it
just leads a mess of local branches.

Instead, let the user specify --branch on the command line or a default
in the git config. The --branch option will override the git config.  If
neither the --branch option is set or the fetchbranch, checkout options
are set, then do not create a branch, which is the same as the
--no-branch option in previous versions.

Keep the --no-branch option, and use it to override the config if it is
set, so this still compatible with old versions.